+++ tests/Integration/HttpLoggingTest.php
@@ -5,17 +5,21 @@ declare(strict_types=1);
 namespace Kreait\Firebase\Tests\Integration;
 
 use Kreait\Firebase\Contract\Auth;
+use Kreait\Firebase\Exception\Auth\UserNotFound;
 use Kreait\Firebase\Tests\IntegrationTestCase;
-use Psr\Log\Test\TestLogger;
+use PHPUnit\Framework\MockObject\MockObject;
+use Psr\Log\LoggerInterface;
 
 /**
  * @internal
  */
 class HttpLoggingTest extends IntegrationTestCase
 {
-    private TestLogger $logger;
+    /** @var MockObject&LoggerInterface */
+    private $logger;
 
-    private TestLogger $debugLogger;
+    /** @var MockObject&LoggerInterface */
+    private $debugLogger;
 
     private Auth $auth;
 
@@ -27,8 +31,8 @@ class HttpLoggingTest extends IntegrationTestCase
     {
         parent::setUp();
 
-        $this->logger = new TestLogger();
-        $this->debugLogger = new TestLogger();
+        $this->logger = $this->createMock(LoggerInterface::class);
+        $this->debugLogger = $this->createMock(LoggerInterface::class);
 
         $this->auth = self::$factory->createAuth();
         $this->authWithLogger = self::$factory->withHttpLogger($this->logger)->createAuth();
@@ -40,8 +44,8 @@ class HttpLoggingTest extends IntegrationTestCase
         $user = $this->auth->createAnonymousUser();
 
         try {
+            $this->logger->expects($this->atLeastOnce())->method('log');
             $this->authWithLogger->getUser($user->uid);
-            $this->assertCount(1, $this->logger->records);
         } finally {
             $this->auth->deleteUser($user->uid);
         }
@@ -49,11 +53,12 @@ class HttpLoggingTest extends IntegrationTestCase
 
     public function testItLogsFailures(): void
     {
+        $this->debugLogger->expects($this->atLeastOnce())->method('log');
+
         try {
             $this->authWithDebugLogger->updateUser('does-not-exist', []);
         } catch (\Throwable $e) {
-        } finally {
-            $this->assertTrue($this->debugLogger->hasNoticeThatContains('USER_NOT_FOUND'));
+            $this->assertInstanceOf(UserNotFound::class, $e);
         }
     }
 
@@ -62,8 +67,8 @@ class HttpLoggingTest extends IntegrationTestCase
         $user = $this->auth->createAnonymousUser();
 
         try {
+            $this->debugLogger->expects($this->atLeastOnce())->method('log');
             $this->authWithDebugLogger->getUser($user->uid);
-            $this->assertCount(1, $this->debugLogger->records);
         } finally {
             $this->auth->deleteUser($user->uid);
         }